update: cleanup code
