+++ backend/internal/utils/utils.go
@@ -2,11 +2,13 @@ package utils
 
 import (
 	"fmt"
-	"github.com/jackc/pgx/v5/pgtype"
 	"log"
 	"math/rand"
 	"os"
+	"strconv"
 	"time"
+
+	"github.com/jackc/pgx/v5/pgtype"
 )
 
 func GetAbsoluteUrl(path string) string {
@@ -39,4 +41,13 @@ func ConvertToPgTimestamp(input string) (pgtype.Timestamp, error) {
 
 	// Assign it to pgtype.Timestamp
 	return pgtype.Timestamp{Time: parsedTime, Valid: true}, nil
+}
+
+func ConvertStringToInt64(input string) int64 {
+	newNum, err := strconv.Atoi(input)
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	return int64(newNum)
 }
