+++ backend/pkg/handlers/parking.go
@@ -2,27 +2,26 @@ package handlers
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
 	"log"
 	"net/http"
-	"strconv"
 	"time"
 
 	db "github.com/careecodes/RentDaddy/internal/db/generated"
+	"github.com/careecodes/RentDaddy/internal/utils"
 	"github.com/clerk/clerk-sdk-go/v2/user"
 	"github.com/go-chi/chi/v5"
 	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-type PermitHandler struct {
+type ParkingPermitHandler struct {
 	pool    *pgxpool.Pool
 	queries *db.Queries
 }
 
-func NewPermitHandler(pool *pgxpool.Pool, queries *db.Queries) *PermitHandler {
-	return &PermitHandler{
+func NewParkingPermitHandler(pool *pgxpool.Pool, queries *db.Queries) *ParkingPermitHandler {
+	return &ParkingPermitHandler{
 		pool,
 		queries,
 	}
@@ -33,16 +32,7 @@ type CreatePermitRequest struct {
 	ExpiresAt    time.Time `json:"expires_at"`
 }
 
-func ConvertStringToInt64(input string) int64 {
-	newNum, err := strconv.Atoi(input)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	return int64(newNum)
-}
-
-func (p PermitHandler) CreateParkingPermitHandler(w http.ResponseWriter, r *http.Request) {
+func (p ParkingPermitHandler) CreateParkingPermitHandler(w http.ResponseWriter, r *http.Request) {
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
 		log.Printf("[PARKING_HANDLER] Failed reading body: %v", err)
@@ -81,19 +71,21 @@ func (p PermitHandler) CreateParkingPermitHandler(w http.ResponseWriter, r *http
 	}
 
 	if count >= 2 {
+		log.Printf("[PARKING_HANDLER] User hit parking permit limit: %d Error: %v", count, err)
 		http.Error(w, "Error parking permit limit reached", http.StatusForbidden)
 		return
 	}
 
 	permit, err := p.queries.CreateParkingPermit(r.Context(), db.CreateParkingPermitParams{
-		PermitNumber: ConvertStringToInt64(req.PermitNumber),
-		CreatedBy:    ConvertStringToInt64(clerkID),
+		PermitNumber: utils.ConvertStringToInt64(req.PermitNumber),
+		CreatedBy:    int64(userMetadata.DbId),
 		ExpiresAt: pgtype.Timestamp{
 			Time:  time.Now().UTC().Add(time.Duration(5) * 24 * time.Hour),
 			Valid: true,
 		},
 	})
 	if err != nil {
+		log.Printf("[PARKING_HANDLER] Failed creating parking permit: %v", err)
 		http.Error(w, "Failed to create parking permit", http.StatusInternalServerError)
 		return
 	}
